backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic and settings
pydantic==2.5.3
pydantic-settings==2.1.0

# MongoDB
motor==3.3.2
pymongo==4.6.1

# PDF extraction
pdfminer.six==20221105
PyPDF2==3.0.1

# Web scraping
newspaper3k==0.2.8
beautifulsoup4==4.12.3
httpx==0.26.0
lxml==5.1.0

# NLP
spacy==3.7.4
rapidfuzz==3.6.1

# Utilities
python-dotenv==1.0.0
